requirements.txt
langchain
faiss-cpu
sentence-transformers
requests
streamlit
ollama
ragas
pandas
numpy
python-dotenv